feat: add method get_business_groups_byuser
Finds business groups that a user belongs to.
They might be filtered by role and/or expanded to take into account SSO/custom groups that the user belongs to.
The returned collection of subtenants contains the list of roles that the user has on those tenants
(without the list of principals that belong to those tenants).